Define SpecificParametersProvider interface
