+++ loadflow/loadflow-api/src/main/java/com/powsybl/loadflow/LoadFlowProvider.java
@@ -9,13 +9,10 @@ package com.powsybl.loadflow;
 
 import com.google.common.collect.Lists;
 import com.powsybl.commons.Versionable;
-import com.powsybl.commons.config.ModuleConfig;
 import com.powsybl.commons.config.PlatformConfig;
 import com.powsybl.commons.config.PlatformConfigNamedProvider;
+import com.powsybl.commons.config.SpecificParametersProvider;
 import com.powsybl.commons.extensions.Extension;
-import com.powsybl.commons.extensions.ExtensionJsonSerializer;
-import com.powsybl.commons.parameters.ConfiguredParameter;
-import com.powsybl.commons.parameters.Parameter;
 import com.powsybl.commons.report.ReportNode;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.iidm.network.Network;
@@ -41,7 +38,7 @@ import java.util.concurrent.CompletableFuture;
  *
  * @author Geoffroy Jamgotchian {@literal <geoffroy.jamgotchian at rte-france.com>}
  */
-public interface LoadFlowProvider extends Versionable, PlatformConfigNamedProvider {
+public interface LoadFlowProvider extends Versionable, PlatformConfigNamedProvider, SpecificParametersProvider<LoadFlowParameters, Extension<LoadFlowParameters>> {
 
     static List<LoadFlowProvider> findAll() {
         return Lists.newArrayList(ServiceLoader.load(LoadFlowProvider.class, LoadFlowProvider.class.getClassLoader()));
@@ -70,33 +67,6 @@ public interface LoadFlowProvider extends Versionable, PlatformConfigNamedProvid
     Optional<Class<? extends Extension<LoadFlowParameters>>> getSpecificParametersClass();
 
     /**
-     * The serializer for implementation-specific parameters, or {@link Optional#empty()} if the implementation
-     * does not have any specific parameters, or does not support JSON serialization.
-     *
-     * <p>Note that the actual serializer type should be {@code ExtensionJsonSerializer<LoadFlowParameters, MyParametersExtension>}
-     * where {@code MyParametersExtension} is the specific parameters class.
-     *
-     * @return The serializer for implementation-specific parameters.
-     */
-    Optional<ExtensionJsonSerializer> getSpecificParametersSerializer();
-
-    /**
-     * Reads implementation-specific parameters from platform config, or return {@link Optional#empty()}
-     * if the implementation does not have any specific parameters, or does not support loading from config.
-     *
-     * @return The specific parameters read from platform config.
-     */
-    Optional<Extension<LoadFlowParameters>> loadSpecificParameters(PlatformConfig config);
-
-    /**
-     * Reads implementation-specific parameters from a Map, or return {@link Optional#empty()}
-     * if the implementation does not have any specific parameters, or does not support loading from config.
-     *
-     * @return The specific parameters read from Map.
-     */
-    Optional<Extension<LoadFlowParameters>> loadSpecificParameters(Map<String, String> properties);
-
-    /**
      * Create a `Map` of parameter name / `String` value from implementation-specific parameters.
      * If the implementation does not have any specific parameters, `Map` is empty.
      *
@@ -106,39 +76,8 @@ public interface LoadFlowProvider extends Versionable, PlatformConfigNamedProvid
     Map<String, String> createMapFromSpecificParameters(Extension<LoadFlowParameters> extension);
 
     /**
-     * Updates implementation-specific parameters from a Map.
-     */
-    void updateSpecificParameters(Extension<LoadFlowParameters> extension, Map<String, String> properties);
-
-    /**
      * Updates implementation-specific parameters from a PlatformConfig
      */
     void updateSpecificParameters(Extension<LoadFlowParameters> extension, PlatformConfig config);
 
-    /**
-     * Get the parameters of the parameters extension associated with this provider.
-     *
-     * @return the parameters of the parameters extension associated with this provider.
-     */
-    List<Parameter> getSpecificParameters();
-
-    /**
-     * Retrieves the parameters of the extension associated with this provider,
-     * incorporating any overrides from the PlatformConfig.
-     *
-     * @return The parameters of the associated extension with overrides applied from PlatformConfig.
-     */
-    default List<Parameter> getSpecificParameters(PlatformConfig platformConfigConfig) {
-        return ConfiguredParameter.load(getSpecificParameters(), getModuleConfig(platformConfigConfig).orElse(null));
-    }
-
-    /**
-     * Retrieves the configuration of the specific module for this provider from the provided PlatformConfig.
-     *
-     * @param platformConfig The platform configuration containing potential module configurations.
-     * @return An Optional containing the ModuleConfig if present, otherwise an empty Optional.
-     */
-    default Optional<ModuleConfig> getModuleConfig(PlatformConfig platformConfig) {
-        throw new UnsupportedOperationException("Not implemented");
-    }
 }
