+++ sensitivity-analysis-api/src/main/java/com/powsybl/sensitivity/SensitivityAnalysisProvider.java
@@ -9,10 +9,9 @@ package com.powsybl.sensitivity;
 
 import com.google.common.collect.Lists;
 import com.powsybl.commons.Versionable;
-import com.powsybl.commons.config.PlatformConfig;
 import com.powsybl.commons.config.PlatformConfigNamedProvider;
+import com.powsybl.commons.config.SpecificParametersProvider;
 import com.powsybl.commons.extensions.Extension;
-import com.powsybl.commons.extensions.ExtensionJsonSerializer;
 import com.powsybl.commons.report.ReportNode;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.contingency.Contingency;
@@ -37,7 +36,7 @@ import java.util.concurrent.CompletableFuture;
  * </p>
  * @author Sebastien Murgey {@literal <sebastien.murgey at rte-france.com>}
  */
-public interface SensitivityAnalysisProvider extends Versionable, PlatformConfigNamedProvider {
+public interface SensitivityAnalysisProvider extends Versionable, PlatformConfigNamedProvider, SpecificParametersProvider<SensitivityAnalysisParameters, Extension<SensitivityAnalysisParameters>> {
 
     static List<SensitivityAnalysisProvider> findAll() {
         return Lists.newArrayList(ServiceLoader.load(SensitivityAnalysisProvider.class, SensitivityAnalysisProvider.class.getClassLoader()));
@@ -69,45 +68,6 @@ public interface SensitivityAnalysisProvider extends Versionable, PlatformConfig
                                 SensitivityAnalysisParameters parameters,
                                 ComputationManager computationManager,
                                 ReportNode reportNode);
-
-    /**
-     * The serializer for implementation-specific parameters, or {@link Optional#empty()} if the implementation
-     * does not have any specific parameters, or does not support JSON serialization.
-     *
-     * <p>Note that the actual serializer type should be {@code ExtensionJsonSerializer<SensitivityAnalysisParameters, MyParametersExtension>}
-     * where {@code MyParametersExtension} is the specific parameters class.
-     *
-     * @return The serializer for implementation-specific parameters.
-     */
-    default Optional<ExtensionJsonSerializer> getSpecificParametersSerializer() {
-        return Optional.empty();
-    }
-
-    /**
-     * Reads implementation-specific parameters from platform config, or return {@link Optional#empty()}
-     * if the implementation does not have any specific parameters, or does not support loading from config.
-     *
-     * @return The specific parameters read from platform config.
-     */
-    default Optional<Extension<SensitivityAnalysisParameters>> loadSpecificParameters(PlatformConfig config) {
-        return Optional.empty();
-    }
-
-    /**
-     * Reads implementation-specific parameters from a Map, or return {@link Optional#empty()}
-     * if the implementation does not have any specific parameters, or does not support loading from config.
-     *
-     * @return The specific parameters read from Map.
-     */
-    default Optional<Extension<SensitivityAnalysisParameters>> loadSpecificParameters(Map<String, String> properties) {
-        return Optional.empty();
-    }
-
-    /**
-     * Updates implementation-specific parameters from a Map.
-     */
-    default void updateSpecificParameters(Extension<SensitivityAnalysisParameters> extension, Map<String, String> properties) {
-    }
 
     /**
      *
